+++ src/main/java/umc/lightup/common/enums/Gender.java
@@ -1,0 +1,5 @@
+package umc.lightup.common.enums;
+
+public enum Gender {
+    MALE, FEMALE
+}
