+++ src/main/java/umc/lightup/common/enums/Mbti.java
@@ -1,0 +1,20 @@
+package umc.lightup.common.enums;
+
+public enum Mbti {
+    ISTJ,
+    ISFJ,
+    INFJ,
+    INTJ,
+    ISTP,
+    ISFP,
+    INFP,
+    INTP,
+    ESTP,
+    ESFP,
+    ENFP,
+    ENTP,
+    ESTJ,
+    ESFJ,
+    ENFJ,
+    ENTJ
+}
