+++ src/main/java/umc/lightup/common/enums/NotificationType.java
@@ -1,0 +1,5 @@
+package umc.lightup.common.enums;
+
+public enum NotificationType {
+    MESSAGE, INVITE, ALERT // 실제 타입에 맞게 추가
+}
