+++ src/main/java/umc/lightup/common/enums/ReferenceType.java
@@ -1,0 +1,5 @@
+package umc.lightup.common.enums;
+
+public enum ReferenceType {
+    ITEM, MEMBER, PROJECT // 실제 타입에 맞게 추가
+}
