+++ src/main/java/umc/lightup/common/enums/Role.java
@@ -1,0 +1,7 @@
+package umc.lightup.common.enums;
+
+public enum Role {
+    LEADER, // 팀원을 구하는 유저
+    TEAMMATE // 팀을 구하는 유저
+}
+
