+++ src/main/java/umc/lightup/common/enums/ToolType.java
@@ -1,0 +1,5 @@
+package umc.lightup.common.enums;
+
+public enum ToolType {
+    SOFTWARE, HARDWARE // 실제 타입에 맞게 추가
+}
