--- src/main/java/umc/lightup/domain/BaseEntity.java
@@ -1,21 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import java.time.LocalDateTime;
-
-@MappedSuperclass
-@Getter
-public abstract class BaseEntity {
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate
-    @Column(nullable = false, updatable = true)
-    private LocalDateTime updatedAt;
-}
