--- src/main/java/umc/lightup/domain/Gender.java
@@ -1,5 +1,0 @@
-package umc.lightup.domain;
-
-public enum Gender {
-    MALE, FEMALE
-}
