--- src/main/java/umc/lightup/domain/ItemLikeGroup.java
@@ -1,17 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-public class ItemLikeGroup extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @Column(length = 20, nullable = false)
-    private String name;
-}
