--- src/main/java/umc/lightup/domain/ItemViewHistory.java
@@ -1,25 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.*;
-import org.springframework.data.annotation.LastModifiedDate;
-
-import java.time.LocalDateTime;
-
-@Entity
-public class ItemViewHistory { //extends 없음
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "item_id", nullable = false)
-    private Item item;
-
-    @LastModifiedDate
-    @Column(nullable = false)
-    private LocalDateTime viewedAt;
-}
