--- src/main/java/umc/lightup/domain/Mbti.java
@@ -1,20 +1,0 @@
-package umc.lightup.domain;
-
-public enum Mbti {
-    ISTJ,
-    ISFJ,
-    INFJ,
-    INTJ,
-    ISTP,
-    ISFP,
-    INFP,
-    INTP,
-    ESTP,
-    ESFP,
-    ENFP,
-    ENTP,
-    ESTJ,
-    ESFJ,
-    ENFJ,
-    ENTJ
-}
