--- src/main/java/umc/lightup/domain/MemberSkill.java
@@ -1,18 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-public class MemberSkill extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "skill_id", nullable = false)
-    private Skill skill;
-}
