--- src/main/java/umc/lightup/domain/Notification.java
@@ -1,35 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-public class Notification extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "sender_id", nullable = false)
-    private Member sender;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "receiver_id", nullable = false)
-    private Member receiver;
-
-    @Enumerated(EnumType.STRING)
-    @Column(nullable = false, length = 30)
-    private NotificationType type;
-
-    @Column(nullable = false, columnDefinition = "TEXT")
-    private String message;
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "reference_type", nullable = false, length = 30)
-    private ReferenceType referenceType;
-
-    @Column(name = "reference_id", nullable = false)
-    private Long referenceId;
-
-    @Column(name = "is_read", nullable = false)
-    private Boolean isRead;
-}
