--- src/main/java/umc/lightup/domain/NotificationType.java
@@ -1,5 +1,0 @@
-package umc.lightup.domain;
-
-public enum NotificationType {
-    MESSAGE, INVITE, ALERT // 실제 타입에 맞게 추가
-}
