--- src/main/java/umc/lightup/domain/Portfolio.java
@@ -1,20 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-public class Portfolio extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "member_id", nullable = false)
-    private Member member;
-
-    @Column(length = 30)
-    private String name;
-
-    @Column(nullable = false, length = 255)
-    private String fileUrl;
-}
