--- src/main/java/umc/lightup/domain/Position.java
@@ -1,16 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-
-@Entity
-public class Position extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Column(length = 30, nullable = false, unique = true)
-    private String name;
-}
