--- src/main/java/umc/lightup/domain/RecruitPosition.java
@@ -1,18 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-public class RecruitPosition extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "item_id", nullable = false)
-    private Item item;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "position_id", nullable = false)
-    private Position position;
-}
