--- src/main/java/umc/lightup/domain/ReferenceType.java
@@ -1,5 +1,0 @@
-package umc.lightup.domain;
-
-public enum ReferenceType {
-    ITEM, MEMBER, PROJECT // 실제 타입에 맞게 추가
-}
