--- src/main/java/umc/lightup/domain/Region.java
@@ -1,22 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.Id;
-
-@Entity
-public class Region extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Column(nullable = false, length = 20)
-    private String sido;
-
-    @Column(nullable = false, length = 30)
-    private String sigungu;
-
-    @Column(unique = true, length = 50)
-    private String fullName;
-}
