--- src/main/java/umc/lightup/domain/Role.java
@@ -1,7 +1,0 @@
-package umc.lightup.domain;
-
-public enum Role {
-    LEADER, // 팀원을 구하는 유저
-    TEAMMATE // 팀을 구하는 유저
-}
-
