--- src/main/java/umc/lightup/domain/Tool.java
@@ -1,17 +1,0 @@
-package umc.lightup.domain;
-
-import jakarta.persistence.*;
-
-@Entity
-public class Tool extends BaseEntity {
-    @Id
-    @GeneratedValue
-    private Long id;
-
-    @Column(length = 30, nullable = false)
-    private String name;
-
-    @Enumerated(EnumType.STRING)
-    @Column(length = 30, nullable = false)
-    private ToolType type;
-}
