--- src/main/java/umc/lightup/domain/ToolType.java
@@ -1,5 +1,0 @@
-package umc.lightup.domain;
-
-public enum ToolType {
-    SOFTWARE, HARDWARE // 실제 타입에 맞게 추가
-}
