+++ src/main/java/umc/lightup/item/domain/ItemLikeGroup.java
@@ -1,0 +1,19 @@
+package umc.lightup.item.domain;
+
+import jakarta.persistence.*;
+import umc.lightup.common.BaseEntity;
+import umc.lightup.member.domain.Member;
+
+@Entity
+public class ItemLikeGroup extends BaseEntity {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @Column(length = 20, nullable = false)
+    private String name;
+}
