+++ src/main/java/umc/lightup/member/domain/MemberStrength.java
@@ -1,0 +1,20 @@
+package umc.lightup.member.domain;
+
+import jakarta.persistence.*;
+import umc.lightup.common.BaseEntity;
+import umc.lightup.strength.domain.Strength;
+
+@Entity
+public class MemberStrength extends BaseEntity {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "member_id", nullable = false)
+    private Member member;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "strength_id", nullable = false)
+    private Strength strength;
+}
