+++ src/main/java/umc/lightup/tool/domain/Tool.java
@@ -1,0 +1,19 @@
+package umc.lightup.tool.domain;
+
+import jakarta.persistence.*;
+import umc.lightup.common.BaseEntity;
+import umc.lightup.common.enums.ToolType;
+
+@Entity
+public class Tool extends BaseEntity {
+    @Id
+    @GeneratedValue
+    private Long id;
+
+    @Column(length = 30, nullable = false)
+    private String name;
+
+    @Enumerated(EnumType.STRING)
+    @Column(length = 30, nullable = false)
+    private ToolType type;
+}
